fix(phase-6a): water_network deepcopy, checkpoints cross-wiring, noise dtype, CI gate
- terrain_pipeline.py: deepcopy water_network at both save and restore in
  _save_checkpoint/rollback_to — prevents aliasing corruption on repeated rollbacks
- terrain_checkpoints.py: external save_checkpoint now populates
  water_network_snapshot (deepcopy), side_effects_snapshot, pass_history_len —
  fixes rollback-through-external-checkpoint wiping pass_history + water_network
- _terrain_noise.py: drop .astype(np.float32) from noise2_array — returns float64
  matching _PermTableNoise for cross-env content-hash determinism
- callable_census_gate.py: utf-8-sig BOM stripping + dunder method skip filter
- callable_census_gate.lock: establish CI ratchet baseline (161 uncovered / 1279,
  87.4% graded)

+++ veilbreakers_terrain/handlers/_terrain_noise.py
@@ -178,7 +178,7 @@ class _OpenSimplexWrapper(_PermTableNoise):
 
     def noise2_array(self, xs: np.ndarray, ys: np.ndarray) -> np.ndarray:
         vf = np.vectorize(self._os.noise2)
-        return vf(xs.ravel(), ys.ravel()).reshape(xs.shape).astype(np.float32)
+        return vf(xs.ravel(), ys.ravel()).reshape(xs.shape)
 
 
 # Legacy alias so that any code importing ``OpenSimplex`` from this module
